pakwachPOS/requirements.txt
acme==1.1.0
aioredis==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
channels==3.0.5
channels-redis==3.3.1
chardet==3.0.4
charset-normalizer==3.2.0
ConfigArgParse==0.13.0
configobj==5.0.6
constantly==15.1.0
cryptography==39.0.1
daphne==3.0.2
distlib==0.3.8
distro==1.4.0
Django==3.2.16
django-cors-headers==3.11.0
django-redis==5.2.0
django-user-agents==0.4.0
django-webpack-loader==2.0.1
djangorestframework==3.12.4
filelock==3.15.4
forex-python==1.8
future==0.18.3
hiredis==2.2.1
HTMLParser==0.0.2
httplib2==0.22.0
hyperlink==21.0.0
idna==2.8
incremental==22.10.0
josepy==1.2.0
meson==1.2.1
mock==3.0.5
more-itertools==10.3.0
msgpack==1.0.4
mysqlclient==2.1.1
netifaces==0.10.4
numpy==1.26.4
oauth2==1.9.0.post1
packaging==24.1
parsedatetime==2.4
pbr==5.4.5
Pillow==9.2.0
platformdirs==4.2.2
pluggy==0.13.1
ply==3.11
protobuf==4.21.12
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
-e git+https://github.com/base4sistemas/pyescpos@e0d435789542b4fb640c6f1b994136f01fffc99a#egg=PyESCPOS
pyOpenSSL==23.0.0
pyparsing==3.1.0
pyRFC3339==1.1
pytest==4.6.11
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==5.3.1
rave-python==1.2.16
redis==4.5.1
requests==2.22.0
requests-toolbelt==0.8.0
rsa==4.9
# Editable install with no version control (sample-python-calculator==1.0.0)
-e /Users/maliabupatricia/Downloads/devskiller-code-G723-HQRG-QRWH-S2F
service-identity==23.1.0
simplejson==3.19.2
six==1.14.0
sqlparse==0.4.3
tox==3.28.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
ua-parser==0.18.0
urllib3==1.25.8
user-agents==2.2.0
virtualenv==20.26.3
wcwidth==0.2.13
zope.component==4.3.0
zope.event==4.4
zope.interface==6.0
